feat(downloads): render the full view inside the title-bar popup
- Move the DownloadsModal design into `DownloadsFullView`, rewired from
  the panel store / window.api to the popup's prop + `__comfyTitlePopup`
  bridge, and replace BaseModal chrome with the popup card
- Drop the per-row entrance animation: its lingering transform layer
  blocked pointer hit-testing in the transparent popup view
- Wire the `downloads-full` branch + card chrome into TitlePopupApp
